+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Full support for user-uploaded files (`/files` endpoints)
   ([#377](https://github.com/Open-EO/openeo-python-client/issues/377))
+- Initial, experimental "local processing" feature to use
+  openEO Python Client Library functionality on local
+  GeoTIFF/NetCDF files and also do the processing locally
+  using the `openeo_processes_dask` package
+  ([#338](https://github.com/Open-EO/openeo-python-client/pull/338))
 
 ### Changed